fix: add support for starknet hash for `getHash`
