+++ src/verify/index.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, describe, vi } from 'vitest';
-import { getHash, verify } from '.';
+import { verify } from '.';
 import evmMessage from '../../test/fixtures/evm/message-alias.json';
 import starknetMessage from '../../test/fixtures/starknet/message-alias.json';
 import * as evmVerification from './evm';
@@ -11,15 +11,6 @@ evmVerificationMock.mockImplementation(() => Promise.resolve(true));
 starknetVerificationMock.mockImplementation(() => Promise.resolve(true));
 
 describe('sign/utils', () => {
-  describe('getHash', () => {
-    test('should return a hash string', () => {
-      const hash = getHash(evmMessage.data);
-      expect(hash).toBe(
-        '0x82ed8be33f43c86f9b83d14736e5762c89108fbc9b8b54f6e993818fc8a53525'
-      );
-    });
-  });
-
   describe('verify', () => {
     test('should call EVM verification on EVM address', async () => {
       expect.assertions(1);
